refactor: add project_id to get_collection and pin_collection requests
This will bring these requests in line with all other requests needing to specify the parent
resource e.g. collection for objects / groups and now project for collections

BREAKING CHANGE: This will change the behaviour in some language stubs
project_id will be needed for these requests in the future. In the transition
phase these will be optional.